requirements.txt
fastapi
uvicorn
torch
torchvision
mlflow
tqdm
Pillow
python-multipart